requirements.txt
numpy>=1.23
PyYAML>=6.0
pydantic>=2.0
rasterio>=1.3
affine>=2.4
matplotlib>=3.4
brokenaxes>=0.6
bmipy>=2.0
geopandas>=0.14
fiona>=1.9
shapely>=2.0
pyproj>=3.4
pywatemsedem
richdem>=1.0
netCDF4>=1.6
